Reinstating PETSC_DIR in package build
Change purely in packaging system: with the move to 3.6.3 on the standard Ubuntu LTS build for Fluidity, the debian packaging method needs PETSC_DIR setting again. For builds NOT on the standard Fluidity Ubuntu LTS platform, PETSC_DIR may need to be reset to whatever local value is required, but since most builds will take place on the standard Fluidity builders this should, in most cases, work.